ATR list command added

+++ src/availableSupportedCards.cpp
@@ -45,10 +45,10 @@ namespace electronic_id
 constexpr char hexmap[] = {'0', '1', '2', '3', '4', '5', '6', '7',
                            '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
 
-std::string hexStr(const unsigned char *data, int len)
+std::string hexStr(const unsigned char *data, size_t len)
 {
     std::string s(len * 2, ' ');
-    for (int i = 0; i < len; ++i) {
+    for (size_t i = 0; i < len; ++i) {
         s[2 * i]     = hexmap[(data[i] & 0xF0) >> 4];
         s[2 * i + 1] = hexmap[data[i] & 0x0F];
     }
